+++ clients/fides-js/src/docs/fides-options.ts
@@ -176,4 +176,28 @@ export interface FidesOptions {
    * Defaults to `undefined`.
    */
   fides_consent_override: "accept" | "reject";
+
+  /**
+   * Given a OneTrust → Fides notice mapping exists and the OneTrust cookie exists, Fides will “migrate” those consents to Fides privacy notices, and write to the Fides cookie.
+   *
+   * This way, Fides customers that are migrating away from OneTrust don’t need to show their users new consent dialogues when switching to Fides.
+   * that those preferences are respected.
+   *
+   * Example original otFidesMapping data:
+   * {
+   *    'C0001': ['essential_cookies'],
+   *    'C0002': ['analytics_tracking'],
+   *    'C0004': ['advertising', 'targeted_ads']
+   * }
+   *
+   * To encode original data to the format expected by this field, use:
+   * encodeURIComponent(JSON.stringify(otFidesMapping))
+   *
+   * To decode this field, use:
+   * JSON.parse(decodeURIComponent(ot_fides_mapping))
+   *
+   * Field defaults to `undefined`.
+   *
+   */
+  ot_fides_mapping: string;
 }